hw1/requirements.txt
numpy
scipy
sklearn
matplotlib
pytorch
